+++ RELEASES.md
@@ -1,9 +1,15 @@
+## v0.2-a (2020-02-28)
+- Bug fixes
+- Implement basic form of StageGroup
+- Pause a Constraint/Stage
+- Remove Stage from StageGroup
 ## v0.1-a (2020-02-27)
+
 - **Initial version of Constraint/Stage system**
+
 - Create a constraint
 - Create combined constraints
 - Create a model
 - Create a stage
 - Abort a constraint
-- Pause a constraint/stage
 - Add a constraint to stage